Update documentation
Make documentation of ELECTRODE fixes more complete by listing more warnings and describing options more fully.
Use utils::logical for toggle (on/off) options.
With the changes in etypes to auto-detect electrode types it makes more sense to make it an on/off toggle as well, so that we don't have inconsistent keyword types.